+++ src/base/config.inc
@@ -39,11 +39,11 @@ CONFIG_CONSTRAINT(CONFIG_KEY(bucket_name).size() > 0, "bucket_name must be speci
   CONFIG(std::string, gs_token_file, "", "path to token file generated by __PACKAGE_NAME___gs_get_token(1); file must not be group- or world-readable or -writable");
 #endif
 
-#ifdef WITH_FVS
-  CONFIG_SECTION("FV/S");
-  CONFIG(std::string, fvs_secret_file, "", "path to file containing FV/S Access Key ID followed by a space followed by FV/S Secret Access Key; file must not be group- or world-readable or -writable");
-  CONFIG(std::string, fvs_service_endpoint, "gss.iijgio.com", "service endpoint for IIJ FV/S");
-  CONFIG(bool, fvs_use_ssl, true, "set to 'no'/'false' to disable SSL");
+#ifdef WITH_IIJGIO
+  CONFIG_SECTION("IIJGIO");
+  CONFIG(std::string, iijgio_secret_file, "", "path to file containing IIJ GIO storage and analysis Access Key ID followed by a space followed by IIJ GIO storage and analysis Secret Access Key; file must not be group- or world-readable or -writable");
+  CONFIG(std::string, iijgio_service_endpoint, "storage-dag.iijgio.com", "service endpoint for IIJ GIO storage and analysis");
+  CONFIG(bool, iijgio_use_ssl, true, "set to 'no'/'false' to disable SSL");
 #endif
 
 CONFIG_SECTION("Encryption");
