--- src/services/fvs/Makefile.am
@@ -1,7 +1,0 @@
-noinst_LIBRARIES = libs3fuse_services_fvs.a
-
-libs3fuse_services_fvs_a_SOURCES = \
-	file_transfer.cc \
-	file_transfer.h \
-	impl.cc \
-	impl.h
