+++ src/services/iijgio/Makefile.am
@@ -1,0 +1,7 @@
+noinst_LIBRARIES = libs3fuse_services_iijgio.a
+
+libs3fuse_services_iijgio_a_SOURCES = \
+	file_transfer.cc \
+	file_transfer.h \
+	impl.cc \
+	impl.h
